Add hook to check Safenet allowlist
