+++ apps/web/src/features/myAccounts/index.tsx
@@ -5,14 +5,13 @@ import { DataWidget } from '@/features/myAccounts/components/DataWidget'
 import { useAllSafesGrouped, type AllSafeItemsGrouped } from '@/features/myAccounts/hooks/useAllSafesGrouped'
 import useTrackSafesCount from '@/features/myAccounts/hooks/useTrackedSafesCount'
 import css from '@/features/myAccounts/styles.module.css'
-import useHasSafenetFeature from '@/features/safenet/hooks/useHasSafenetFeature'
+import useIsWalletSafenetAllowlisted from '@/features/safenet/hooks/useIsWalletSafenetAllowlisted'
 import useWallet from '@/hooks/wallets/useWallet'
-import { useDeploySafenetAccountQuery } from '@/store/safenet'
 import madProps from '@/utils/mad-props'
 import { Box, Divider, Paper } from '@mui/material'
 import classNames from 'classnames'
 import dynamic from 'next/dynamic'
-import { useMemo, useState } from 'react'
+import { useState } from 'react'
 
 const SafenetBanner = dynamic(() => import('@/features/safenet/components/new-safe/SafenetBanner'))
 
@@ -27,28 +26,14 @@ const MyAccounts = ({ safes, onLinkClick, isSidebar = false }: MyAccountsProps) 
   const [searchQuery, setSearchQuery] = useState('')
   useTrackSafesCount(safes, wallet)
 
-  const safeDeploymentProps = useMemo(() => {
-    return {
-      account: {
-        owners: wallet?.address ? [wallet.address] : [],
-        threshold: 1,
-      },
-      saltNonce: Date.now().toString(),
-      dryRun: true,
-    }
-  }, [wallet?.address])
-
-  const hasSafenetFeature = useHasSafenetFeature()
-  const { status: safenetDeployDryRunStatus } = useDeploySafenetAccountQuery(safeDeploymentProps, {
-    skip: !hasSafenetFeature,
-  })
+  const canDeploySafenetAccount = useIsWalletSafenetAllowlisted()
 
   return (
     <Box data-testid="sidebar-safe-container" className={css.container}>
       <Box className={classNames(css.myAccounts, { [css.sidebarAccounts]: isSidebar })}>
         <AccountsHeader isSidebar={isSidebar} onLinkClick={onLinkClick} />
 
-        {!isSidebar && safenetDeployDryRunStatus === 'fulfilled' && <SafenetBanner />}
+        {!isSidebar && canDeploySafenetAccount && <SafenetBanner />}
 
         <Paper sx={{ padding: 0 }}>
           <AccountListFilters setSearchQuery={setSearchQuery} />
